# Shader define expression optimization