@Dok11 2023-06-08
> I am not a huge fan of the static way the cache is working. basically it only grows to 50000 and then no cache. I wonder if there would be a way to not keep oldest entries or rotate them instead ?

Of course there are many things that we can make with memoization cache, but the current value seems a huge limit and this property is not private, so it can be changed, if developer would like to increase cache size.
I can make cache more smart if you want.