@sebavan 2023-06-08
would be cool as long as it does not slow things down @Popov72 any thoughts ?