@Dok11 2023-06-09
To save performance we should remove outdated cache pretty rarely, not each cache access.
So, as an idea, in the cache we will save the last time access, when we reach the limit we will run cache clear that removes half of cache stack (25,000 in out case) by picking items that a long time was not accessed.
As other way, we can save the number of access times to each cache item and while cache clear removes only items that rarely needed.
I can try some of this idea, if you like it. Or let's discuss what else we can make here.