@sebavan 2023-06-09
Let s ensure the cache write access  for the date won t slow things more than it was before the PR