@Dok11 2023-06-10
> Let s ensure the cache write access for the date won t slow things more than it was before the PR

If InfixToPostfixCacheLimitSize have not enough size and clearCache calls every time, than it has slow down. For example, my scene utilize 168 cache items while loading and if InfixToPostfixCacheLimitSize=50 (just fifty items) and every fiftieth half items in cache removes that it really slow down to 27.86ms for all function calls and 15.37ms for first 10 000 calls - it is better than before PR.

But when InfixToPostfixCacheLimitSize has enough size I measured 5.57ms for first 10 000 calls and 12.13ms for all calls.